sync components/merge (extracting from components/operations.py)
