+++ tests/unit/test_api.py
@@ -16,13 +16,12 @@ class ApiTest:
     def test_can_reach_module_alias_symbols_from_base(self):
         import datumaro as dm
 
-        assert hasattr(dm.ops, "ExactMerge")
         assert hasattr(dm.project, "Project")
         assert hasattr(dm.errors, "DatumaroError")
 
     @mark_requirement(Requirements.DATUM_API)
     def test_can_import_from_module_aliases(self):
         # pylint: disable=unused-import
+        from datumaro.components.merge.exact_merge import ExactMerge
         from datumaro.errors import DatumaroError
-        from datumaro.ops import ExactMerge
         from datumaro.project import Project
