+++ tests/unit/test_ops.py
@@ -19,13 +19,15 @@ from datumaro.components.annotation import (
 from datumaro.components.annotations import match_segments_pair
 from datumaro.components.dataset import Dataset
 from datumaro.components.dataset_base import DatasetItem
-from datumaro.components.media import Image, MultiframeImage, PointCloud
-from datumaro.components.operations import (
+from datumaro.components.errors import (
     FailedAttrVotingError,
-    IntersectMerge,
     NoMatchingAnnError,
     NoMatchingItemError,
     WrongGroupError,
+)
+from datumaro.components.media import Image, MultiframeImage, PointCloud
+from datumaro.components.merge.intersect_merge import IntersectMerge
+from datumaro.components.operations import (
     compute_ann_statistics,
     compute_image_statistics,
     find_unique_images,
@@ -557,7 +559,7 @@ class TestMultimerge(TestCase):
         )
 
         merger = IntersectMerge()
-        merged = merger([source0, source1, source2])
+        merged = merger(source0, source1, source2)
 
         compare_datasets(self, expected, merged)
         self.assertEqual(
@@ -710,7 +712,7 @@ class TestMultimerge(TestCase):
         )
 
         merger = IntersectMerge(conf={"quorum": 1, "pairwise_dist": 0.1})
-        merged = merger([source0, source1, source2])
+        merged = merger(source0, source1, source2)
 
         compare_datasets(self, expected, merged, ignored_attrs={"score"})
         self.assertEqual(
@@ -769,7 +771,7 @@ class TestMultimerge(TestCase):
         )
 
         merger = IntersectMerge(conf={"quorum": 1, "pairwise_dist": 0.1})
-        merged = merger([source0, source1])
+        merged = merger(source0, source1)
 
         compare_datasets(self, expected, merged, ignored_attrs={"score"})
         self.assertEqual(0, len(merger.errors))
@@ -846,7 +848,7 @@ class TestMultimerge(TestCase):
         )
 
         merger = IntersectMerge(conf={"quorum": 3, "ignored_attributes": {"ignored"}})
-        merged = merger([source0, source1, source2])
+        merged = merger(source0, source1, source2)
 
         compare_datasets(self, expected, merged, ignored_attrs={"score"})
         self.assertEqual(2, len([e for e in merger.errors if isinstance(e, FailedAttrVotingError)]))
@@ -873,7 +875,7 @@ class TestMultimerge(TestCase):
         )
 
         merger = IntersectMerge(conf={"groups": [["a", "a_g1", "a_g2_opt?"], ["c", "c_g1_opt?"]]})
-        merger([dataset, dataset])
+        merger(dataset, dataset)
 
         self.assertEqual(
             3, len([e for e in merger.errors if isinstance(e, WrongGroupError)]), merger.errors
@@ -927,7 +929,7 @@ class TestMultimerge(TestCase):
         )
 
         merger = IntersectMerge()
-        merged = merger([source0, source1])
+        merged = merger(source0, source1)
 
         compare_datasets(self, expected, merged, ignored_attrs={"score"})
 
@@ -1015,7 +1017,7 @@ class TestMultimerge(TestCase):
         )
 
         merger = IntersectMerge()
-        merged = merger([source0, source1])
+        merged = merger(source0, source1)
 
         compare_datasets(self, expected, merged, ignored_attrs={"score"})
 
@@ -1069,7 +1071,7 @@ class TestMultimerge(TestCase):
         )
 
         merger = IntersectMerge()
-        merged = merger([source0, source1])
+        merged = merger(source0, source1)
 
         compare_datasets(self, expected, merged)
 
@@ -1112,6 +1114,6 @@ class TestMultimerge(TestCase):
         )
 
         merger = IntersectMerge()
-        merged = merger([source0, source1])
+        merged = merger(source0, source1)
 
         compare_datasets(self, expected, merged)
